les: create costtable file
